+++ NEWS.md
@@ -2,7 +2,8 @@
 
 # Version 0.1.1
 
-- `SampledStructureFunction` renamed as `EmpiricalStructureFunction`.
+- `SampledStructureFunction` renamed as `EmpiricalStructureFunction` and
+  `ShrinkedLazyCovariance` renamed as `PackedLazyCovariance`.
 
 - Lazy covariance and empirical structure function object have (read-only)
   public properties to favor using the `A.f` syntax.